Add Association and UserAssociation services
Integrated new services for managing Associations and UserAssociations. The AssociationService contains methods for retrieving and creating associations specific to a user. The UserAssociationService is responsible for creating user-association relationships. Additionally, adjustments were made to the UserAuthProvider and SecurityConfig for supporting these new services. The schema for GraphQL was updated to include the 'getMyAssociations' query.