Add DefaultImage model, repository and implementation
Created new DefaultImage model and repository to store default images for the system. Added these default images to the database during startup of the ShootingClubManagerApplication. User accounts now have a default image attached when registered through the AuthController. Furthermore, adjusted the expiration time and updated the SECRET in JwtService for improved security.