Implement reservation management features
Added functionality to manage reservations in the application. Code allows for creation, modification, and deletion of individual as well as series of reservations. Modifications to data transfer objects and other related classes were made to support these changes. Additionally, the system of repeating reservations was simplified by removing the 'CUSTOM' repeat type option.