Implement custom authentication token and refactor error messages
Added a CustomUsernamePasswordAuthenticationToken class to allow access to the client's IP address in the authentication token. Mutations have been migrated to GraphQL and major refactoring of the responses was performed in GlobalRequestExceptionHandler. Also, error messages handling was simplified to provide more concise output to the clients. In AuthController, controller annotation was changed to be compatible with GraphQL.