Refactor code and add email service
This code update removes unnecessary import statements, performs dependency adjustments, and removes TestController. In addition, the email service for user registration has also been added in the application. Also, there are adjustments to user authentication, changes in the way user input is validated in AuthController and the addition of an HTML email sending service.