Add scheduled task to mark expired competitions as inactive
Implemented a new scheduled task, CompetitionScheduledTask, which runs every six hours to check for competitions that have ended and marks them as inactive. Complemented this with changes in the AssociationCompetitionService and AssociationCompetitionRepository to fetch the list of competitions that need to be marked inactive. Also added 'active' field in Competition model. Enabled scheduling in the main application with the @EnableScheduling annotation.